refactored package structure: created parser subdir in directory internal
